Projects list and template gallery become separate pages
